Fix: use parseFloat instead of parseInt for quote unit prices
parseInt truncates decimals (e.g. "0.5" → 0), causing sub-dollar
amounts to fail totaling. Changed all 8 occurrences across frontend
(new + edit pages) and backend (POST + PATCH APIs).

+++ app/admin/quote-system/[id]/edit/page.tsx
@@ -108,7 +108,7 @@ export default function EditQuotePage() {
   };
 
   const subtotal = items.reduce(
-    (sum, item) => sum + (parseInt(item.unitPrice) || 0) * (parseInt(item.quantity) || 0),
+    (sum, item) => sum + (parseFloat(item.unitPrice) || 0) * (parseFloat(item.quantity) || 0),
     0
   );
   const discountAmount = subtotal * (parseFloat(form.discount) || 0) / 100;
@@ -272,7 +272,7 @@ export default function EditQuotePage() {
                   </div>
                   <div className="flex items-center">
                     <span className="text-sm text-gray-500 whitespace-nowrap">
-                      ${((parseInt(item.unitPrice) || 0) * (parseInt(item.quantity) || 0)).toLocaleString()}
+                      ${((parseFloat(item.unitPrice) || 0) * (parseFloat(item.quantity) || 0)).toLocaleString()}
                     </span>
                   </div>
                 </div>
